// app/src/types/device.ts
export interface Device {
  tikTokHandle: string;
  url: string;
  airDroidLink: string;
  image: string;
  serialNumber: string;
  deviceNumber: string;
  platform: string;
  model: string;
  client: string;
  clientEmail: string;
  phoneNumber: string;
  mobileCarrier: string;
  imei: string;
  notes: string;
  checked: string;
  status: string;
  timestamp: string;
  airDroidId: string;
  password: string;
} 